// src/sports/enums/supported-league.enum.ts
export enum SupportedLeague {
  PREMIER_LEAGUE = 'PREMIER_LEAGUE',
  CHAMPIONSHIP = 'CHAMPIONSHIP',

  LA_LIGA = 'LA_LIGA',
  SERIE_A = 'SERIE_A',
  BUNDESLIGA = 'BUNDESLIGA',
  LIGUE_1 = 'LIGUE_1',

  EREDIVISIE = 'EREDIVISIE',
  PRIMEIRA_LIGA = 'PRIMEIRA_LIGA',
  SCOTTISH_PREMIERSHIP = 'SCOTTISH_PREMIERSHIP',
  BELGIAN_PRO_LEAGUE = 'BELGIAN_PRO_LEAGUE',
  TURKISH_SUPER_LIG = 'TURKISH_SUPER_LIG',

  NPFL = 'NPFL',
  SOUTH_AFRICA_PREMIER_DIVISION = 'SOUTH_AFRICA_PREMIER_DIVISION',
  EGYPTIAN_PREMIER_LEAGUE = 'EGYPTIAN_PREMIER_LEAGUE',
  BOTOLA_PRO = 'BOTOLA_PRO',

  BRAZIL_SERIE_A = 'BRAZIL_SERIE_A',
  ARGENTINA_PRIMERA = 'ARGENTINA_PRIMERA',
  LIGA_MX = 'LIGA_MX',
  MLS = 'MLS',
  SAUDI_PRO_LEAGUE = 'SAUDI_PRO_LEAGUE',
}