// src/sports/enums/supported-club-competition.enum.ts
export enum SupportedClubCompetition {
  UEFA_CHAMPIONS_LEAGUE = 'UEFA_CHAMPIONS_LEAGUE',
  UEFA_EUROPA_LEAGUE = 'UEFA_EUROPA_LEAGUE',
  UEFA_CONFERENCE_LEAGUE = 'UEFA_CONFERENCE_LEAGUE',

  CAF_CHAMPIONS_LEAGUE = 'CAF_CHAMPIONS_LEAGUE',
  CAF_CONFEDERATION_CUP = 'CAF_CONFEDERATION_CUP',

  COPA_LIBERTADORES = 'COPA_LIBERTADORES',
  COPA_SUDAMERICANA = 'COPA_SUDAMERICANA',

  CONCACAF_CHAMPIONS_CUP = 'CONCACAF_CHAMPIONS_CUP',
}