// src/sports/enums/supported-international-competition.enum.ts
export enum SupportedInternationalCompetition {
  FIFA_WORLD_CUP = 'FIFA_WORLD_CUP',
  FIFA_WORLD_CUP_QUALIFIERS = 'FIFA_WORLD_CUP_QUALIFIERS',

  AFCON = 'AFCON',
  AFCON_QUALIFIERS = 'AFCON_QUALIFIERS',

  UEFA_EURO = 'UEFA_EURO',
  UEFA_EURO_QUALIFIERS = 'UEFA_EURO_QUALIFIERS',
  UEFA_NATIONS_LEAGUE = 'UEFA_NATIONS_LEAGUE',

  COPA_AMERICA = 'COPA_AMERICA',

  CONCACAF_GOLD_CUP = 'CONCACAF_GOLD_CUP',
  CONCACAF_NATIONS_LEAGUE = 'CONCACAF_NATIONS_LEAGUE',

  AFC_ASIAN_CUP = 'AFC_ASIAN_CUP',
}